requirements.txt
streamlit
lancedb
lance-namespace
sentence-transformers
torch
torchvision
ollama
pillow
requests
beautifulsoup4
sqlalchemy
tqdm
deprecation